Move the FileawayError into a separate file.
